app/requirements.txt
appdirs==1.4.4
Babel==2.8.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
chart-studio==1.1.0
click==7.1.2
configparser==4.0.2
contextlib2==0.6.0.post1
cryptography==2.9.2
distlib==0.3.0
enum34==1.1.10
filelock==3.0.12
Flask==1.1.2
Flask-Babel==1.0.0
Flask-Table==0.5.0
Flask-WTF==0.14.3
idna==2.9
importlib-metadata==1.6.0
importlib-resources==1.5.0
ipaddress==1.0.23
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.20
numpy==1.16.6
pandas==0.24.2
pathlib2==2.3.5
plotly==4.7.1
pycparser==2.20
PyMySQL==0.9.3
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
retrying==1.3.3
scandir==1.10.0
singledispatch==3.4.0.3
six==1.14.0
typing==3.7.4.1
urllib3==1.25.9
virtualenv==20.0.20
Werkzeug==1.0.1
WTForms==2.3.1
zipp==1.2.0
